Create a HashMap to hold all the datatypes
Each datatype is a type of data that will be collected on every
interval.

Datatype has a struct Data member whose implementation needs only
contain the data collected for time t. All details of saving the data
collected to a file, timing constraints are handled in the DataType
implementation.

This is in preparation for adding data types for collection.