Use TimeEnum for storing time data
Use TimeEnum for time instead of DateTime<Utc>. This will allow us to
reuse the field to hold absolute time difference.

This is in preparation for adding support for the visualizer, where the
data returned will be a diff of values at time t and time t+1. The time
difference will be stored in the same time field set with the
TimeEnum::TimeDiff type.

Also, fix Diskstats to have the right timestamp for each data collected.
It was set to use the starting time always.