Add support for visualizing Kernel Config data
Implement the logic to pull data from the Kernel Config files directly. This
is the same as the implementation in the procfs crate. We need access to
the CONFIGs that are commented out as well as the headers. Additionally,
kernel_config() returns a HashMap which mixes up the order of the values.