Use ProcessedData for the visualizer
There is a new method added to GetData i.e. process_raw_data. This will be passed the raw
data gathered by the collector. This will return a Vec<ProcessedData> which will be held
in memory and be used to get_data() when the API requests show up.

The collector will now store the data it gets from procfs as is to the YAML file. It will
not parse it and store it as a struct. This helps to reduce the overhead of collection.

Change the visualizer to use ProcessedData instead of Data. Also change the tests file
to reflect the new way of gathering data.