Use the run name as dir name
The test name specified by the user will be the name of the directory
where the performance data is stored. If no name is given, a default
aperf_<timestamp> is used.

Additionally, when running the visualizer, if two directories are given
to be compared, the names cannot be the same. Throw an error in such a
case and exit.

Finally, change the binary name to aperf. This involves changes to the
release.yaml file and the logging variable (APERF_LOG_LEVEL).