Explicitly use sysconf to get total CPUs
The num_cpus crate checks for schedaffinity and queries for the total
number of CPUs in the system using CPU_SETSIZE. This will cause the cpu
count to be incorrect in systems where CGROUPS and schedaffinity rules
are applied to aperf. Use sysconf to query for the total number of
online CPUs in the system.