Gracefully fail when unable to find run data
If the directory or archive fed to aperf report does not contain valid
run data, fail gracefully. If the total number of fails is equal to the
total number of visualizers then the run data is invalid. If only some
of the visualizers fail during init, it could be that some data was not
collected. This will help in being backwards compatible.