Allow for creating reports from other reports
The APerf reports contain the data archives. Use these if a report is
passed in to the aperf report sub-command.

The aperf report command will fail if a report is given as an arg and
the report name is omitted.

Also, move to using PathBuf instead of String for path.