Create a test harness
For integration testing, we need a test harness which can setup the temp
directory. All the files generated as part of the tests should be
located in this dir. The test harness will be able to catch assert
failures and re-raise them after cleaning up the temp directory.

This requires all integration tests run as closures in run_test().