Add support for configuring the X axis
The Configure tab will show the 'Aggregate CPU Utilization' graph for
all runs. The user can use the graph along with a rangeslider to
configure the X-Axis for all the graphs in the report.

Configuration can be carried out either in the graph or the rangeslider.
Double-clicking on the main graph resets the X-Axis to default.