Aperf: Allow to use custom PMU
Allow the user to specify a custom PMU file to be used during aperf
record. This custom PMU file must be in the expected format.

To aid in the creation of the custom PMU file, add a sub-command.

'aperf custom-pmu' - an interactive sub-command to create custom PMU.
This allows the user to create a file from scratch or modify an existing
PMU config aperf already knows about. The custom file generated can be
used by specifying '--pmu-config' with 'aperf record'.

Additionally, a pmu-config file can be verified by specifying --verify
along with --pmu-file with 'aperf custom-pmu'.