Report: Fix issue with '.' in run names
There is a bug in the report generation if the user collected record
data with a run name containing '.'. In such cases, the report
generator, since it was using file_stem(), would incorrectly create the
report directory and tarball with different names.

To prevent this, use file_name() and a check to remove '.tar.gz' suffix.
Also, move to using file_name() instead of file_stem() wherever needed.

Add an integration test to check with a run name containing a '.'.